backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
python-multipart